requirements.txt
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
kiwisolver==1.4.7
lxml==5.3.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
plum-dispatch==1.7.4
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==21.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
Spire.Doc==12.7.1
typing_extensions==4.12.2
urllib3==2.2.2
